// src/apis/generated/model/todoDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * BFF API
 * Frontend 向け BFF API
 * OpenAPI spec version: 1.0
 */

export interface TodoDto {
  /** TODO ID */
  id: string;
  /** TODOタイトル */
  title: string;
  /** 完了状態 */
  completed: boolean;
  /** 作成日時。ISO 8601 date-time 文字列 */
  createdAt: string;
}